refactor: share sleep helpers across startup and store
